+++ godot-core/src/meta/mod.rs
@@ -60,6 +60,9 @@ pub use class_name::ClassName;
 pub use godot_convert::{FromGodot, GodotConvert, ToGodot};
 pub use traits::{ArrayElement, GodotType, PackedArrayElement};
 
+#[cfg(since_api = "4.2")]
+pub use crate::registry::signal::variadic::ParamTuple;
+
 pub(crate) use array_type_info::ArrayTypeInfo;
 pub(crate) use traits::{
     element_godot_type_name, element_variant_type, GodotFfiVariant, GodotNullableFfi,
