+++ godot-core/src/registry/mod.rs
@@ -16,11 +16,11 @@ pub mod plugin;
 pub mod property;
 
 #[cfg(since_api = "4.2")]
-pub mod functional;
+pub mod signal;
 
 // Contents re-exported in `godot` crate; just keep empty.
 #[cfg(before_api = "4.2")]
-pub mod functional {}
+pub mod signal {}
 
 // RpcConfig uses MultiplayerPeer::TransferMode and MultiplayerApi::RpcMode, which are only enabled in `codegen-full` feature.
 #[cfg(feature = "codegen-full")]
