+++ godot-core/src/registry/signal/mod.rs
@@ -7,8 +7,9 @@
 
 mod connect_builder;
 mod typed_signal;
-mod variadic;
+pub(crate) mod variadic;
 
 pub use connect_builder::*;
 pub use typed_signal::*;
-pub use variadic::*;
+pub use variadic::SignalReceiver;
+// ParamTuple re-exported in crate::meta.
