+++ godot/src/prelude.rs
@@ -32,10 +32,9 @@ pub use super::obj::{
 };
 
 // Make trait methods available.
-pub use super::obj::WithBaseField as _;
-pub use super::obj::WithSignals as _; // signals() // base(), base_mut(), to_gd()
-
 pub use super::obj::EngineBitfield as _;
 pub use super::obj::EngineEnum as _;
 pub use super::obj::NewAlloc as _;
 pub use super::obj::NewGd as _;
+pub use super::obj::WithBaseField as _; // base(), base_mut(), to_gd()
+pub use super::obj::WithSignals as _; // signals()
